requirements.txt
fsspec==2024.2.0
gcsfs==2024.2.0
google-cloud-aiplatform==1.41.0
google-cloud-storage==2.14.0
grpc-google-iam-v1==0.13.0
joblib==1.3.2
kfp==2.7.0
kfp-pipeline-spec==0.3.0
kfp-server-api==2.0.5
numpy==1.24.4
packaging==23.2
pandas==2.0.3
protobuf==4.25.2
pyarrow==15.0.0
requests==2.31.0
scikit-learn==1.3.2
pytest==8.0.1


